+++ app/Http/Controllers/HousingBController.php
@@ -14,11 +14,17 @@ class HousingBController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Punishment $student_id)
     {
         $stud = HousingB::get();
-        $puns = Punishment::latest()->first();
-        return view('pages.super.housing.housingB', ['stud' => $stud, 'puns' => $puns]);
+        $puns = Punishment::whereIn('student_id', $stud->pluck('student_id'))->latest()->get()->groupBy('student_id');
+
+        // $puns = Punishment::get();
+        // $puns = Punishment::whereIn('student_id', $stud->pluck('student_id'))->latest()->get();
+        // $pun_one = Punishment::where('student_id',$student_id)->latest()->first();
+
+        $stud_one=[];
+        return view('pages.super.housing.housingB', ['stud' => $stud, 'puns' => $puns,'stud_one'=>$stud_one ]);
     }
     
 
@@ -36,6 +42,7 @@ class HousingBController extends Controller
      */
     public function store(Request $request)
     {
+
         //inster student 
         $stud = new HousingB();
         // $stud->id
@@ -60,9 +67,8 @@ class HousingBController extends Controller
         $stud->date_joined = $request->date_joined;
         $stud->dorm_joined = $dorm_joined;
         $stud->phone = $request->phone;
-        // $stud->punishment_id = Hash::make($request->password); 
-        // $stud->punishment_id = $request->punishment_id;
         $stud->save();
+
 
 
         return redirect('housingB');
@@ -71,31 +77,34 @@ class HousingBController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(HousingB $student)
+    public function show( $id)
     {
-        //
+        $stud_one = HousingB::where('id',$id)->get();
+        $stud = HousingB::get();
+        $puns = Punishment::get();
+        return view('pages.super.housing.housingB', ['stud' => $stud, 'puns' => $puns ,'stud_one'=>$stud_one]);
+
     }
 
     /**
      * Show the form for editing the specified resource.
      */
     public function edit($id){
+        // dd($id);
                 $stud=HousingB::find($id);
                 if(! $stud){
         
                     return redirect('housingB');
                 }
-                    
                 return view('pages.super.housing.updateStudentB', compact('stud'));
-        
             }
 
     /**
      * Update the specified resource in storage.
      */
+    // public function update(Request $request, Student $student)
     public function update(Request $request, $id)
     {
-       
         $stud = HousingB::findOrFail($id);
         $stud->student_id = $request->student_id;
         $stud->name = $request->name;
@@ -104,6 +113,7 @@ class HousingBController extends Controller
         $stud->date_joined = $request->date_joined;
         $stud->phone = $request->phone;
         $stud->save();
+
         return redirect('housingB');
 
 
@@ -114,6 +124,7 @@ class HousingBController extends Controller
      * Remove the specified resource from storage.
      */
 
+
         public function destroy($student_id)
 {
     $stud = HousingB::where('student_id', $student_id)->first();
@@ -121,14 +132,20 @@ class HousingBController extends Controller
     $stud2= Punishment::where('student_id',$student_id)->first();
 
 
-    // if($stud && $stud->user_id == Student::user()->id){
-         $stud->delete();
-         $stud1->delete();  
-         $stud2->delete();     
-        return back()->with('success','student  deleted successfully' );
-    // }else{
-        // abort(404);
+
+    if ($stud) {
+        $stud->delete();
+    }
+
+    if ($stud1) {
+        $stud1->delete();
+    }
+
+    if ($stud2) {
+        $stud2->delete();
+    }
+
+        return redirect()->back();
     }
 }
-    // }
